+++ core/sr-api-macros/tests/ui/missing_block_generic_parameter.stderr
@@ -1,0 +1,13 @@
+error: Missing `Block` generic parameter.
+  --> $DIR/missing_block_generic_parameter.rs:19:13
+   |
+19 |     impl self::Api for Runtime {
+   |                ^^^
+
+error[E0107]: wrong number of type arguments: expected 1, found 0
+  --> $DIR/missing_block_generic_parameter.rs:19:7
+   |
+19 |     impl self::Api for Runtime {
+   |          ^^^^^^^^^ expected 1 type argument
+
+For more information about this error, try `rustc --explain E0107`.
